use byte[] to save varint length table

+++ src/main/java/me/steinborn/krypton/mod/shared/network/util/VarIntUtil.java
@@ -5,11 +5,11 @@ package me.steinborn.krypton.mod.shared.network.util;
  * from zero to 32.
  */
 public class VarIntUtil {
-    private static final int[] VARINT_EXACT_BYTE_LENGTHS = new int[33];
+    private static final byte[] VARINT_EXACT_BYTE_LENGTHS = new byte[33];
 
     static {
         for (int i = 0; i <= 32; ++i) {
-            VARINT_EXACT_BYTE_LENGTHS[i] = (int) Math.ceil((31d - (i - 1)) / 7d);
+            VARINT_EXACT_BYTE_LENGTHS[i] = (byte) ((int) Math.ceil((31d - (i - 1)) / 7d));
         }
         VARINT_EXACT_BYTE_LENGTHS[32] = 1; // Special case for 0.
     }
